NaviCube: Improved label rendering
This change should fix the unreadable text on Windows by removing the
QFont's stretch option.

Labels are now set to upeercase not before but after the translation
to not break any ongoing translation work.

Label's font size is now soft-contrained for better readability.
It means that if label does not fit into the cube's face it will
be adjusted automatically. One can notice that the bottom face of
the cube now uses slightly smaller font than the other faces.