UI/Preferences: Realign the Preferences UI
Width of the list of preference groups is being set via function call,
while still having dimensions set in the *.ui file. These two places
can and actually are currently diverged. This results in bad looking UI.

This change adjusts the Preferences list width to the widest item
once it is fully populated with list items. It also reduces the list item
spacing which makes the list a bit slimmer and neat.